Support content type headers for akka-http
Previously the content type was always set to text, which could cause
issues for marshallers that enforce the content type is set to
application/json. This issue did not effect http4s, as the "body" field
for a request is just a stream of bytes, with no assigned content type.